VPAAMP-118 Remove memcpy from demux sample extraction
