+++ middleware/InterfacePlayerRDK.h
@@ -37,6 +37,7 @@
 #include "SocUtils.h"
 #include "GstUtils.h"
 #include "DemuxDataTypes.h"
+#include "MediaSample.h"
 
 // Forward declarations
 class InterfacePlayerPriv;
